ItemFastAPI/requirements.txt
pymongo==3.12.0
fastapi==0.66.0
uvicorn==0.14.0
pydantic==1.8.2
typing-extensions==3.7.4.3
python-decouple==3.4
python-multipart
amqp
appdirs
billiard
celery
certifi
click
click-didyoumean
click-plugins
click-repl
distlib
filelock
h11
install
kombu
pipenv
pippip
prompt-toolkit
pytz
redis
six
starlette
vine
virtualenv
virtualenv-clone
wcwidth
yapf
rabbitmq

